# Reduce memory needed for string allocation tracing