This PR reduces the memory used creating duplicate strings especially in the case of very long strings which require a lot of memory.

This code builds on the string cache from #51 and @NickLaMuro's md5 ideas in #48 to only store the 200 characters needed for the String Report in the output. The `object_id` of the cached, shorted string summary works like the md5 hash to uniquely identify string allocations even when strings are the same in the first 200 characters.

I reorganzined the `string_report` method to be a little more readable (hopefully) while also introducting the object_id logic. The switch to using end-of-line dot notation makes it possible to have inline comments but I could change that back.

@NickLaMuro I've tested this with the ManageIQ example and it now profiles completely on startup!