@dgynn 2017-10-05
Great feedback, @NickLaMuro. I wasn't thinking about the fact that allocating more strings while building the string_report is a big problem because we are still in a GC disabled state on the first pass. I'd normally prefer to sort strings than arrays but not in this case.

After bringing in your diff to group_by using arrays (of existing objects) I saw this could be optimized more by just grouping by the string_value.object_id and using the `list[0].string_value` trick to get the string back out of the list.

I think this shaved off about 25% of the time in `string_report` or about 5% of the total time and should keep memory low. Can you run your test again and see if this looks right?

One more change I'm considering would be to try to take the "top" before doing the sort. The change below speeds up `string_report` by about 90% (in this ManageIQ case) but could produce inconsistent results when strings are allocated the same number of times (possibly rare).
```diff
diff --git a/lib/memory_profiler/results.rb b/lib/memory_profiler/results.rb
index b675150..bc37e6a 100644
--- a/lib/memory_profiler/results.rb
+++ b/lib/memory_profiler/results.rb
@@ -51,8 +51,8 @@ module MemoryProfiler
         keep_if { |stat| stat.string_value }.
         group_by { |stat| stat.string_value.object_id }.
         values.
+        max(top) { |a, b| a.size <=> b.size }.
         sort_by! { |list| [-list.size, list[0].string_value] }.
-        first(top).
         # Return array of [string, [[location, count], [location, count], ...]
         map! { |list| [list[0].string_value,
                        list.group_by { |stat| stat.location }.
```