@NickLaMuro 2017-10-05
> Great feedback, @NickLaMuro.

Cool, glad it wasn't pedantic!  Though, I guess my standards are allowed to be a little higher on a performance tool PR review ;)

> I wasn't thinking about the fact that allocating more strings while building the string_report is a big problem because we are still in a GC disabled state on the first pass. I'd normally prefer to sort strings than arrays but not in this case.

So this part might not be quite accurate, but it is possible I am still getting up to speed code base.

But if I am not mistaken, the `GC` is enabled [here](https://github.com/SamSaffron/memory_profiler/blob/a75c677/lib/memory_profiler/reporter.rb#L47), and the `results.rb` code in question happens after that, so we aren't going to be growing without `GC` available to help things out.  That said, doesn't hurt to cut down on memory usage at this point when possible.

> After bringing in your diff to group_by using arrays (of existing objects) I saw this could be optimized more by just grouping by the `string_value.object_id` and using the `list[0].string_value` trick to get the string back out of the list.

Nice, good catch.  I keep forgetting that the uniq `object_ids` now give us a pretty solid uniq key to work with.

> Can you run your test again and see if this looks right?

Absolutely.  Will do.

> One more change I'm considering would be to try to take the "top" before doing the sort...

Hmmm, I wonder if multiplying the `top` value by 1.5 (or something) and use that value for the `.max` call, then keep the `.first(top)` call as is, and that reduce the inaccuracy rate of that in most cases, but keep most of the performance gains in place.  Thoughts?