@NickLaMuro 2017-10-05
So here are the benchmarks, but they were a little surprising when I applied your patch from the diff in your previous comment:

**Current**

```
Run 1:  time => 66.90s, mem => 486.3MB (peak ~501MB)
Run 2:  time => 67.35s, mem => 491.4MB (peak... missed it...)
Run 3:  time => 67.26s, mem => 491.1MB (peak 503.9MB)
```

**After Patch**

```
Run 1:  time => 68.41s, mem => 510.8MB
Run 2:  time => 65.10s, mem => 513.4MB
Run 3:  time => 66.04s, mem => 506.7MB
```

Saw no overall speed improvements after applying your patch (didn't test just the `#string_report` part though), but the extra memory was the surprising bit, and makes me question if I did my tests wrong, or applied the diff incorrectly.  Just looking at what you suggested makes me thing that this would provide a benefit across the board, but what I measured says otherwise.

Maybe you would have better luck?  I can try again later as well.