@dgynn 2017-10-09
@NickLaMuro You are totally right. GC has been re-enabled by that point. Thanks for pointing that out.

In this ManageIQ test case that we are looking at (621,863 strings allocated, 155,669 unique) ~70% of the time is spent in `MemoryProfiler::Helpers#lookup_string` only ~10-12% in `MemoryProfiler::Results#string_report`. So tuning `string_report` isn't going to move the needle very much overall. Still, no point in not optimizing it. The last commit I pushed gets that down to ~5% by reducing the number of arrays that get to the `sort_by!`.

Here is a little script I run to generate ruby-prof reports. Add `gem 'ruby-prof'` to the  Gemfile.dev.rb before running.
```ruby
profile = "profile-#{Time.now.strftime('%F-%H%M%S')}"
report = nil
result = RubyProf.profile { report = MemoryProfiler.report { _("foo") } }
RubyProf::MultiPrinter.new(result).print(profile: profile, threshold: 0)
report.pretty_print
puts "Generated ruby-prof reports as #{profile}.*.html"
```

So with this, we should get the reduced memory of using already-allocated objects, create and sort fewer arrays, and maintain the results ordering by size and string.

Let me know if you see any difference with this latest commit. I'm not sure why you would have seen a memory increase in the last run other than the fact that memory is hard to measure.