@dgynn 2017-10-19
Thanks, @NickLaMuro. I agree, I think this is ready to merge. I just added one very minor tweak to do the sort without sorting arrays.

@SamSaffron Does this look good to merge to you? It could be squashed as well.