@dgynn 2017-12-04
@SamSaffron This PR should be good to merge now. It handles the large-string memory explosion in the cache but still retains the unique string tracking. Let me know if you have any questions or concerns.