@dgynn 2018-07-18
Yep. Sorry for sitting on this so long. I stalled trying to think of cases where the change from using String concatenation to String substring (with `[]`) could create an issue. That really shouldn't happen unless some object reports that its class is `String` but hasn't fully implemented all the methods.