@dgynn 2018-07-21
I just pushed up one more commit that sorts the locations within the string report. It was failing intermittently and now should be stable.